# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ICAMMDA Website API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all posts
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: published
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostBody"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /partners:
    get:
      operationId: listPartners
      tags: [partners]
      summary: List all partners
      responses:
        "200":
          description: List of partners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Partner"
    post:
      operationId: createPartner
      tags: [partners]
      summary: Create a partner
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartnerBody"
      responses:
        "201":
          description: Created partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
  /partners/{id}:
    patch:
      operationId: updatePartner
      tags: [partners]
      summary: Update a partner
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePartnerBody"
      responses:
        "200":
          description: Updated partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
        "404":
          description: Not found
    delete:
      operationId: deletePartner
      tags: [partners]
      summary: Delete a partner
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        published:
          type: boolean
        featured:
          type: boolean
        authorName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - slug
        - content
        - category
        - published
        - featured
        - createdAt
        - updatedAt
    CreatePostBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        published:
          type: boolean
        featured:
          type: boolean
        authorName:
          type: ["string", "null"]
      required:
        - title
        - slug
        - content
        - category
    UpdatePostBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        published:
          type: boolean
        featured:
          type: boolean
        authorName:
          type: ["string", "null"]
...
    Partner:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - order
        - createdAt
        - updatedAt
    CreatePartnerBody:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        order:
          type: integer
      required:
        - name
    UpdatePartnerBody:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        order:
          type: integer
...